+++ lib/components/link.vue
@@ -13,8 +13,7 @@
        :event="event"
        :tag="tag"
        :class="linkClassObject"
-       @click="linkClick"
-    >
+       @click="linkClick">
         <slot></slot>
     </a>
 
@@ -26,16 +25,15 @@
        :target="target || null"
        :rel="computedRel"
        :class="linkClassObject"
-       @click="linkClick"
-    >
+       @click="linkClick">
         <slot></slot>
     </a>
 </template>
 
 <script>
-    import linkMixin from '../mixins/link';
+import linkMixin from '../mixins/link';
 
-    export default {
-        mixins: [linkMixin]
-    };
+export default {
+    mixins: [linkMixin]
+};
 </script>
