+++ lib/mixins/link-base.js
@@ -1,4 +1,4 @@
-export const props = {
+const linkProps = {
     active: {
         type: Boolean,
         default: false
@@ -65,35 +65,13 @@ export const props = {
     }
 };
 
-export const bLinkProps = function() {
-    return Object.keys(props).reduce((memo, prop) => {
-        return Object.assign(memo, this[prop]);
-    }, {});
-};
-
 export default {
-    props,
+    props: linkProps,
     computed: {
-        bLinkProps
+        bLinkProps() {
+            return Object.keys(linkProps).reduce((memo, prop) => {
+                return Object.assign(memo, this[prop]);
+            }, {})
+        }
     }
-};
-
-export const pickProps = (...propsToPick) => {
-    return propsToPick.reduce((memo, prop) => {
-        if (props[prop]) {
-            Object.assign(memo, { [prop]: props[prop] });
-        }
-
-        return memo;
-    }, {});
-};
-
-export const omitProps = (...propsToOmit) => {
-    return Object.keys(props).reduce((memo, prop) => {
-        if (!propsToOmit.includes(prop)) {
-            Object.assign(memo, { [prop]: props[prop] });
-        }
-
-        return memo;
-    }, {});
 };
