+++ lib/mixins/link.js
@@ -1,12 +1,18 @@
 // Props compatible with vue-router
 // https://github.com/vuejs/vue-router/blob/dev/src/components/link.js
-import { props } from "./link-base-props";
+import linkBase from "./link-base";
+const { props, computed } = linkBase;
 
 export default {
+    props,
+
     computed: {
+        ...computed,
+
         isRouterLink() {
             return Boolean(this.$router && this.to && !this.disabled);
         },
+
         _href() {
             if (this.disabled) {
                 return "#";
@@ -22,12 +28,14 @@ export default {
                 return this.to;
             }
         },
+
         computedRel() {
             if (this.target === "_blank" && this.rel === null) {
                 return "noopener";
             }
             return this.rel || null;
         },
+
         linkClassObject() {
             return [
                 this.active ? (this.exact ? this.exactActiveClass : this.activeClass) : null,
@@ -35,8 +43,6 @@ export default {
             ];
         }
     },
-
-    props,
 
     methods: {
         linkClick(e) {
